+++ R/get_built_files.R
@@ -5,16 +5,26 @@ get_built_files <- function(lesson = ".") {
   } else {
     path <- lesson$path
   }
-  if (!fs::dir_exists(fs::path(path, "site", "built"))) { 
+  if (!fs::dir_exists(fs::path(path, "site", "built"))) {
     txt <- "No files built. Run {.code sandpaper::build_lesson()} to build."
     cli::cli_alert_warning(txt)
     return(NULL)
-  } 
+  }
   lfiles <- fs::path_file(lesson$files)
-  built_files <- fs::path(path, "site", "built", fs::path_ext_set(lfiles, "md"))
-  res <- purrr::map(built_files, 
-    ~Episode$new(.x, process_tags = FALSE, fix_liquid = FALSE, fix_links = FALSE)$confirm_sandpaper())
+  built_dir <- fs::path(path, "site", "built")
+  built_files <- fs::dir_ls(built_dir, glob = "*.md")
+  res <- purrr::map(
+    built_files,
+    function(f) {
+      Episode$new(f,
+        process_tags = FALSE,
+        fix_liquid = FALSE,
+        fix_links = FALSE
+      )$confirm_sandpaper()
+    }
+  )
 
   names(res) <- fs::path_rel(built_files, path)
   res
 }
+